Move split methods into Evaluation.Splits
This commit also creates a file for each split
method, as well as improves the runtime of each
split method that works using yhreads (leave_k_out
and matrix_split).
Now, instead of creating a new thread for running a
split on a specific user, and then terminate it, now
we use a thread pool to reuse existing threads when
a user split is complete. This reduces the average
runtime of approximately 40%.