Change recommend method to assume that all items might be useful
This change removes the default value required to consider
an item useful for recommendation (which was the model
interaction threshold). This consisted on filtering out, from
all the ranked items, those with a model prediction bellow this
interaction threshold.

Now all ranked items are considered useful. This change was
done to avoid having models weak on assigning relevant prediction
values, but strong on ranking items using a prediction scale
different from the interaction value scale present on the data set.